# api/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Analytics service knows lifecycle of every task.",
        "title": "Analytics service",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/agreed": {
            "get": {
                "description": "Get count of all signed tasks",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "count"
                ],
                "summary": "get signed tasks count",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/canceled": {
            "get": {
                "description": "Get count of all tasks which are not signed (rejected and \"in process\")",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "count"
                ],
                "summary": "get unsigned tasks count",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/total_time": {
            "get": {
                "description": "Get total signition time in seconds of particular task by its id",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "time"
                ],
                "summary": "get signition time of task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "uuid of task",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "time in sec: 100500",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "user input is invalid",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server can't retrieve signition time",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "": {
            "description": "Cookie authorization requires a pair of Access and Refresh tokens. Get them in auth service. The below url is the place where you can do it.",
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "auth-service/v1/auth/validate"
        }
    },
    "x-extension-openapi": {
        "example": "value on a json format"
    }
}